Distinguish between static and dynamic compatibility
static is for judging whether we can make a move, given the current
board type

dynamic is for judging whether a given move might force a board type

We had both checks checking the conditions for the dynamic check, which
means that if (for example) the board is in RightOrMiddle state, the
space (1, 1) is declared as an illegal move. It should be legal (it's
compatible with the static RightOrMiddle state), but it will clearly
induce the board to become Right.